requirements.txt
jsonpath==0.82.2
pytest==7.3.1
pytest-html
PyYAML==6.0
pyyaml-include==1.3.2
requests==2.22.0
allure-pytest
